proxy_server/requirements.txt
bittensor==9.10.0
fastapi
numpy==2.3.2
protobuf
pydantic==2.11.7
pytest==8.1.1
python-dotenv>=1.0.0
python-dateutil>=2.8.0
redis==5.0.1
Requests==2.32.5
soundfile==0.13.1
uvicorn==0.35.0
wandb
python-multipart
boto3
botocore
librosa
httpx
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0